Add page to view courses
