+++ client/package.json
@@ -12,6 +12,8 @@
     "@ant-design/icons": "^5.3.0",
     "@expo/metro-runtime": "~3.1.1",
     "@react-native-community/slider": "^4.4.2",
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/native-stack": "^6.9.17",
     "axios": "^1.6.7",
     "expo": "~50.0.5",
     "expo-status-bar": "~1.11.1",
@@ -24,7 +26,9 @@
     "react-native-vector-icons": "^10.0.3",
     "react-native-web": "~0.19.6",
     "react-query": "^3.39.3",
-    "reanimated-color-picker": "^3.0.2"
+    "reanimated-color-picker": "^3.0.2",
+    "react-native-screens": "~3.29.0",
+    "react-native-safe-area-context": "4.8.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
